requirements.txt
streamlit>=1.56
plotly>=5.24
pandas>=2.1
openpyxl>=3.1
anthropic>=0.40
requests>=2.31
tenacity>=8.2
python-dotenv>=1.0
